Moving dummy data functions into generic data functions library
